web: break out ConfigEditor from SettingsLabgrid into own file
This allows us to use the ConfigEditor on other screens as well.
